Fix double click EMPTY piece in hand bug
The function isWhite() returns true if the square is empty.
This causes a problem where white is allowed to take empty
squares' index into their hand.

This commit adds a check for this and returns to intended
behaviour: white can only take indicies that contain white
pieces into the hand.